finish implemetation

+++ src/settings.rs
@@ -11,9 +11,9 @@ pub struct Settings {
 #[derive(Debug, Deserialize, Clone)]
 pub struct Discord {
     pub token: String,
-    pub application_id: u32,
+    pub application_id: u64,
     pub guild_id: u32,
-    pub admin_channel_id: u32,
+    pub admin_channel_id: u64,
 }
 
 #[derive(Debug, Deserialize, Clone)]
@@ -29,7 +29,7 @@ pub struct Rule {
 
 #[derive(Debug, Deserialize, Clone)]
 pub struct Channel {
-    pub id: u32,
+    pub id: u64,
     pub keep_images: bool,
     pub max_message_age_in_h: u32,
 }
